Fixed import error (moved `k8s` module inside `control_panel_api`)
Must have created the file outside by accident.
